Fix project structure
- Focus of this commit is to reset the whole project
  to a state where we can continue its development.
  This contains rather minimal changes so that stuff
  actually works if used from other project.
- Making this project to compile and more close
  what other spring cloud projects expect.
- Split code into core, zookeeper and hazelcast modules.
- Added simple tests.
