+++ .gitignore
@@ -9,7 +9,7 @@
 .gradle
 build
 bin
-/target/
+target
 *.swp
 .idea
 *.iml
