+++ docs/src/main/ruby/generate_readme.sh
