+++ spring-cloud-cluster-core/src/main/java/org/springframework/cloud/cluster/leader/AbstractCandidate.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.cloud.cluster.leader;
+
+import java.util.UUID;
+
+/**
+ * Base implementation of a {@link Candidate}.
+ * 
+ * @author Janne Valkealahti
+ *
+ */
+public abstract class AbstractCandidate implements Candidate {
+
+	private static final String DEFAULT_ROLE = "leader";
+	
+	private final String id;
+	
+	private final String role;
+	
+	public AbstractCandidate() {
+		this(UUID.randomUUID().toString(), DEFAULT_ROLE);
+	}
+
+	public AbstractCandidate(String id, String role) {
+		this.id = id;
+		this.role = role;
+	}
+
+	@Override
+	public String getRole() {
+		return this.role;
+	}
+
+	@Override
+	public String getId() {
+		return this.id;
+	}
+
+	@Override
+	public abstract void onGranted(Context ctx) throws InterruptedException;
+
+	@Override
+	public abstract void onRevoked(Context ctx);
+
+}
