+++ spring-cloud-cluster-hazelcast/pom.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>spring-cloud-cluster-hazelcast</artifactId>
+	<packaging>jar</packaging>
+
+	<name>spring-cloud-cluster-hazelcast</name>
+	<description>Spring Cloud Cluster Hazelcast</description>
+
+	<parent>
+		<groupId>org.springframework.cloud</groupId>
+		<artifactId>spring-cloud-cluster</artifactId>
+		<version>1.0.0.BUILD-SNAPSHOT</version>
+		<relativePath>..</relativePath>
+	</parent>
+
+	<dependencies>
+		<dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-cluster-core</artifactId>
+		</dependency>
+		<dependency>
+			<artifactId>hazelcast</artifactId>
+			<groupId>com.hazelcast</groupId>
+		</dependency>
+		<dependency>
+			<artifactId>hazelcast-spring</artifactId>
+			<groupId>com.hazelcast</groupId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+			<scope>test</scope>
+		</dependency>
+	</dependencies>
+
+</project>
