--- src/main/java/org/springframework/cloud/cluster/zk/leader/ZKServer.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2014 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.cloud.cluster.zk.leader;
-
-import org.apache.curator.test.TestingServer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.context.annotation.Bean;
-
-/**
- * @author Patrick Peralta
- */
-public class ZKServer {
-	public static final int ZK_SERVER_PORT = 3121;
-
-	@Bean
-	public TestingServer zkServer() throws Exception {
-		return new TestingServer(ZK_SERVER_PORT);
-	}
-
-	/**
-	 * Main bootstrap method.
-	 */
-	public static void main(String[] args) {
-		SpringApplication.run(ZKTestApplication.class, args);
-	}
-
-}
