--- src/main/resources/application.properties
@@ -1,15 +1,0 @@
-main.type.zk=org.springframework.cloud.cluster.zk.leader.ZKTestApplication
-main.type.zkserver=org.springframework.cloud.cluster.zk.leader.ZKServer
-main.type.hz=org.springframework.cloud.cluster.hazelcast.leader.HazelcastTestApplication
-main.type.=
-
-spring.main.sources=${main.type.${type:}}
-
-# logging options
-logging.level.org.apache.zookeeper=WARN
-logging.level.org.springframework=WARN
-logging.level.org.springframework.cloud=INFO
-
-# shell options
-shell.auth.simple.user.password=letmein
-shell.ssh.port=0
